% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fun_exreport.R
\name{exreport}
\alias{exreport}
\title{Create a new exreport document}
\usage{
exreport(title)
}
\arguments{
\item{title}{A string representing a short title for this document}
}
\value{
an empty exreport document
}
\description{
This function inits a new exreport document to start adding elements for
later rendering.
}
\seealso{
exreportRender, exreportAdd
}

