% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fun_exreport.R
\name{exreportAdd}
\alias{exreportAdd}
\title{Add elements to an existing exreport document}
\usage{
exreportAdd(rep, elem)
}
\arguments{
\item{rep}{a reportable object or a list of them}
}
\value{
an extended exreport document
}
\description{
This function allows to add one or more reportable objects to an exisiting
exreport document.
}
\examples{
# Create an empty document:
report <- exreport("Test document")

# Create a reportable object (an experiment)
experiment <- expCreate(wekaExperiment, name="test-exp", parameter="fold")

# Add this object to the document
exreportAdd(exreport, experiment)
}

