% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_addDoc.R
\docType{data}
\name{ILsExperiment}
\alias{ILsExperiment}
\title{Problem: Comparison between two Ionic Liquids packs to capture and reduce the
emission of CO2 in industrial fuel combustion processes.}
\format{A data frame with the data detailed in the Description.}
\usage{
data(ILsExperiment)
}
\description{
A problem containing the percentage of the CO2 reduction in the emission of 20 industrial
fuel combustion processes. It has been used two different Ionic Liquids (ILs) packs with
different properties. The packs has been reused up to three times, and each experiment
has been repeated three times under the same conditions. The variables of the problem are as follows:
}
\details{
\itemize{
  \item IL The name of the IL packs (IL1 and IL2).
  \item Scenario The name of the industrial fuel combustion process (from Scenario 1 to Scenario20).
  \item Execution The number of the execution for each experiment under the same conditions.
  \item Reused The number of experiments which the IL pack has been used previously (from 0 to 2).
  \item CO2 The percentage of CO2 which has been reduced from the emission.
}
}
\keyword{problems}

