% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_addDoc.R
\docType{data}
\name{wekaExperiment}
\alias{wekaExperiment}
\title{Problem: Comparison between several Machine Learning algorithms from the Weka library}
\format{A data frame with the data detailed in the Description.}
\usage{
data(wekaExperiment)
}
\description{
A problem containing experimental data obtaining by comparing several instances
of Machine Algorithms from the Weka library. The variables are as follows:
}
\details{
\itemize{
  \item method. Classification algorithms used in the experimen (NaiveBayes, J48, IBk)
  \item problem. Problems used as benchmark in the comparison, up to 12.
  \item featureSelection. Boolean parameter indicating if the data was preprocessed
  \item fold. For each configuration a 10-fold cross validation was performed. This variable is a numeric value ranging from 1 to 10.
  \item accuracy. This is a measure of the performance of each algorithm. Representing the percentage of correctly classified instances.
  \item trainingTime. A second measure of performance. This one indicates the time in seconds that took the algorithm to build the model.
}
}
\keyword{problems}

